+++ CHANGELOG.md
@@ -7,6 +7,7 @@ NOTES:
 BUG FIXES:
 * resource/hopsworksai_backup: Handle get backup if backup not found and backupPipeline is InProgress
 * resource/hopsworksai_backup: Wait for cluster start during backup pipeline
+* resource/hopsworksai_backup: Fix interface conversion error 
 
 ENHANCEMENTS: